requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
certifi==2021.5.30
chardet==4.0.0
discord-components==1.1.4
discord.py==1.6.0
idna @ file:///tmp/build/80754af9/idna_1609784788857/work
money==1.3.0
multidict @ file:///C:/ci/multidict_1607349747897/work
packaging==20.9
Pint==0.17
pyparsing==2.4.7
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1611751222202/work
wincertstore==0.2
XlsxWriter @ file:///home/ktietz/src/ci_ipy2/xlsxwriter_1612556818988/work
yarl @ file:///C:/ci/yarl_1606940155993/work
